+++ physics/physics.vcxproj
@@ -50,6 +50,8 @@
     <ClInclude Include="harmonic_damping_body.hpp" />
     <ClInclude Include="integration_parameters.hpp" />
     <ClInclude Include="integration_parameters_body.hpp" />
+    <ClInclude Include="lagrange_equipotentials.hpp" />
+    <ClInclude Include="lagrange_equipotentials_body.hpp" />
     <ClInclude Include="mechanical_system.hpp" />
     <ClInclude Include="mechanical_system_body.hpp" />
     <ClInclude Include="continuous_trajectory_body.hpp" />
