+++ physics/physics.vcxproj.filters
@@ -260,6 +260,12 @@
     <ClInclude Include="tensors.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="lagrange_equipotentials.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="lagrange_equipotentials_body.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="degrees_of_freedom_test.cpp">
